+++ app/components/UI/TemplateRenderer/SafeComponentList.ts
@@ -4,7 +4,7 @@ import SheetHeader from '../../../component-library/components/Sheet/SheetHeader
 import Text from '../../../component-library/components/Texts/Text';
 import Icon from '../../../component-library/components/Icons/Icon';
 import BottomSheetFooter from '../../../component-library/components/BottomSheets/BottomSheetFooter';
-import SmartTransactionStatus from '../../Views/SmartTransactionStatus/smart-transaction-status';
+import SmartTransactionStatus from '../../Views/SmartTransactionStatus/SmartTransactionStatus';
 import { View } from 'react-native';
 
 export const safeComponentList = {
