+++ app/components/Views/SmartTransactionStatus/SmartTransactionStatus.tsx
